requirements.txt
africastalking==1.2.3
alembic==1.5.4
amqp==5.0.5
aniso8601==8.1.1
APScheduler==3.7.0
bcrypt==3.2.0
billiard==3.6.3.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
contextlib2==0.6.0.post1
cryptography==3.4.4
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
gunicorn==20.0.4
idna==3.1
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
phonenumbers==8.12.49
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.27.1
schema==0.7.4
six==1.15.0
SQLAlchemy==1.3.23
tzlocal==2.1
urllib3==1.26.3
vine==5.0.0
Werkzeug==1.0.1
zipp==3.4.0
